Move redux setup to container for now
