+++ app/views/layouts/application.html.erb
@@ -8,7 +8,7 @@
     <%= render 'spree/shared/google_analytics.js' %>
     <%= yield %>
 
-    <div id='root'>
+    <%= react_component 'HelloWorldContainer', { name: 'SOMEONE' }, { prerender: false } %>
     <%= javascript_pack_tag 'application' %>
   </body>
 </html>
