+++ config/application.rb
@@ -44,7 +44,7 @@ module SparkStarterKit
     # React-rails server side renderning
     config.react.server_renderer_pool_size  ||= 1  # ExecJS doesn't allow more than one on MRI
     config.react.server_renderer_timeout    ||= 20 # seconds
-    config.react.server_renderer = React::ServerRendering::SprocketsRenderer
+    # config.react.server_renderer = React::ServerRendering::SprocketsRenderer
 
     config.react.server_renderer_options = {
       code: 'var window = this; ',
