+++ package.json
@@ -31,7 +31,6 @@
     "babel-plugin-transform-object-rest-spread": "^6.26.0",
     "babel-preset-react": "^6.24.1",
     "caniuse-lite": "^1.0.30000885",
-    "eslint-plugin-react": "^7.11.1",
     "mirror-creator": "^1.1.0",
     "prop-types": "^15.6.2",
     "react": "^16.5.1",
@@ -43,6 +42,7 @@
     "webpack-merge": "^4.1.4"
   },
   "devDependencies": {
+    "eslint-plugin-react": "^7.11.1",
     "redux-devtools": "^3.4.1",
     "redux-devtools-dock-monitor": "^1.1.3",
     "redux-devtools-log-monitor": "^1.4.0",
